Make sure to access the decoder with existing keys
Without this fix this Processor behaves differently depending on whether
it is run with DEBUG output or not for ILD like detectors. The reason is
that the cellID encoding string there uses "cell[X|Y]" instead of
"[x|y]". Hence, without this fix an exception is raised in the debug
output if it is used and the rest of the loop is skipped.

This was missed so far in ILD reconstruction because the Processor was
run with Verbosity=SILENT there for quite some time.